requirements.txt
Pattern==2.6
Pillow==3.0.0
beautifulsoup4==4.4.1
cloudpickle==0.1.1
cymem==1.30
decorator==4.0.4
imageio==1.4
moviepy==0.2.2.11
murmurhash==0.24
networkx==1.10
numpy==1.10.1
plac==0.9.1
preshed==0.43
requests==2.8.1
six==1.10.0
spacy==0.99
text-unidecode==1.0
thinc==4.0.0
tqdm==2.2.4
treelib==1.3.2
ujson==1.33
wsgiref==0.1.2
